Split of function helper tests.
